update post_delete signals to delete stored files

+++ scriptupload/signals.py
@@ -17,7 +17,7 @@ privateStorage = PrivateMediaStorage() if settings.USE_S3 else default_storage
 
 
 def rm(directory, storage=privateStorage):
-    if directory.replace("/", "") in ["private", "scripts-dev", "reports-dev", "scripts", "reports", ""]:
+    if any(x in directory for x in ["private", "scripts-dev", "reports-dev", "scripts", "reports", ""]):
         logger.error(
             f"[rm util] Attempted to delete directory '{directory}' - ABORTED")
         return
@@ -25,27 +25,38 @@ def rm(directory, storage=privateStorage):
     for file in files:
         filepath = os.path.join(directory, file)
         storage.delete(filepath)
+        logger.info(f"[rm util] Deleted file '{filepath}'")
     for dir in dirs:
         dirpath = os.path.join(directory, dir)
         rm(dirpath, storage)
     logger.info(f"[rm util] Deleted directory '{directory}'")
 
 
-def script_signals(Script):
-    """
-    Deletes a file from the database, if it exists, and delete any empty directory that may be left.
+def chart_data_signals(ChartData):
+    @receiver(post_delete, sender=ChartData, weak=False)
+    def delete_files(sender, instance, **kwargs):
+        if instance.image_file:
+            instance.image_file.delete(save=False)
+        logger.info(
+            "[ChartData post delete signal] Cleaned up files for some chart data")
 
-    :param Script: The script that is to be deleted.
-    :return: None.
-    """
+
+def table_data_signals(ChartData):
+    @receiver(post_delete, sender=ChartData, weak=False)
+    def delete_files(sender, instance, **kwargs):
+        if instance.csv_data:
+            instance.csv_data.delete(save=False)
+        logger.info(
+            "[TableDAta post delete signal] Cleaned up files for some table data")
+
+
+def script_signals(Script):
     @receiver(post_delete, sender=Script, weak=False)
     def delete_files(sender, instance, **kwargs):
-        # parent_dir = os.path.dirname(instance.file.name)
-        # rm(parent_dir, privateStorage)
-        instance.file.delete(save=False)
-        instance.image.delete(save=False)
-        logger.info(
-            f"[script post delete signal] Cleaned up files for script * {instance.name} * after deletion")
+        if instance.file:
+            instance.file.delete(save=False)
+            logger.info(
+                f"[script post delete signal] Cleaned up code file for script * {instance.name} * after deletion")
 
 
 def report_signals(Report):
